refactor(30-02): extract tenant settings queryOptions
- Create tenant-settings-keys.ts with tenantSettingsQueries
- Move TenantProfile, TenantSettings, TenantNotificationPreferences
- Simplify useTenantSettings/useTenantNotificationPreferences
- Re-export types and queries from use-tenant-settings.ts
